+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [Added Axios interceptor wrapper to be used with `nativeAuth`](https://github.com/ElrondNetwork/dapp-core/pull/506)
 - [Added `nativeAuth` and Storybook](https://github.com/ElrondNetwork/dapp-core/pull/504)
 
 ## [[2.1.20](https://github.com/ElrondNetwork/dapp-core/pull/500)] - 2022-12-07